// shared/enums.ts
export enum PORTAL_COLLECTIONS {
    Portals = 'portals',
}

export enum GP_Portal_Enitities {
    Person = 'person',
    Vehicle = 'vehicle',
    All = 'all',
}

export enum PORTAL_GATE_INTERACTIONS {
    SWITCH = 'portal-gate:Switch',
    SHOW_MENU = 'portal-gate:Open:Menu',
    TOGGLE_LOCK = 'portal-gate:Toggle:Lock',
    ENTER = 'portal-gate:Enter',
    PORT = 'portal-gate:Port',
    REMOVE = 'portal-gate:Remove',
    SET_NAME = 'portal-gate:SetName',
    SET_PORTAL_NAME = 'portal:SetName',
}

export enum PORTAL_IPL_CONFIG {
    AMMUNATIONS_GUN_STORE = 'ammunationsGunStore',
    AMMUNATIONS_GUN_CLUB = 'ammunationsGunClub',
    APARTMENT1 = 'apartment1',
    APARTMENT10 = 'apartment10',
    APARTMENT11 = 'apartment11',
    APARTMENT2 = 'apartment2',
    APARTMENT3 = 'apartment3',
    APARTMENT4 = 'apartment4',
    APARTMENT5 = 'apartment5',
    APARTMENT6 = 'apartment6',
    APARTMENT7 = 'apartment7',
    APARTMENT8 = 'apartment8',
    APARTMENT9 = 'apartment9',
    BAHAMAMAMAS = 'bahamamamas',
    BASE = 'base',
    BIKERCLUBHOUSE1 = 'bikerclubhouse1',
    BIKERCLUBHOUSE2 = 'bikerclubhouse2',
    BILLBOARDS = 'billboards',
    BLIPS = 'blips',
    BUNKERS = 'bunkers',
    CAYOPERICO = 'cayoperico',
    COCAINE = 'cocaine',
    COUNTERFEIT = 'counterfeit',
    DOCUMENT = 'document',
    DRUGWARS = 'drugwars',
    FACILITY = 'facility',
    FLOYD = 'floyd',
    FRANKLIN = 'franklin',
    FRANKLINAUNT = 'franklinaunt',
    FREAKSHOW = 'freakshow',
    GARAGE = 'garage',
    GARAGE1 = 'garage1',
    GARAGE2 = 'garage2',
    GARAGE3 = 'garage3',
    GARAGE4 = 'garage4',
    GARAGE5 = 'garage5',
    GRAFFITIS = 'graffitis',
    GUNYACHT = 'gunyacht',
    HANGAR = 'hangar',
    HEISTCARRIER = 'heistcarrier',
    HOUSE1 = 'house1',
    HOUSE10 = 'house10',
    HOUSE2 = 'house2',
    HOUSE3 = 'house3',
    HOUSE4 = 'house4',
    HOUSE5 = 'house5',
    HOUSE6 = 'house6',
    HOUSE7 = 'house7',
    HOUSE8 = 'house8',
    HOUSE9 = 'house9',
    INDEX = 'index',
    LESTERFACTORY = 'lesterfactory',
    MEETUP = 'meetup',
    METH = 'meth',
    METHLAB = 'methlab',
    MICHAEL = 'michael',
    MUSicrooftop = 'musicrooftop',
    NEWMORGUE = 'newmorgue',
    NIGHTCLUB = 'nightclub',
    NORTHYANKTON = 'northyankton',
    OFFICE1 = 'office1',
    OFFICE2 = 'office2',
    OFFICE3 = 'office3',
    OFFICE4 = 'office4',
    OFFICE5 = 'office5',
    OFFICE6 = 'office6',
    OFFICE7 = 'office7',
    OFFICE8 = 'office8',
    REDCARPET = 'redcarpet',
    RVVAN = 'rvvan',
    SIMEON = 'simeon',
    STRIPCLUB = 'stripclub',
    STUDIO = 'studio',
    TRAINCRASH = 'traincrash',
    TREVORSTRAILER = 'trevorstrailer',
    TUNER = 'tuner',
    UFOS = 'ufos',
    UNLOAD = 'unload',
    VEHICLEWAREHOUSE = 'vehiclewarehouse',
    WAREHOUSE = 'warehouse',
    WEED = 'weed',
    YACHT = 'yacht',
    ZANCUDOGATES = 'zancudogates',
}